Move set_date to utils module.
