+++ src/util/__init__.py
@@ -1,6 +1,5 @@
 """The util library."""
 from .input import get_input, get_number, get_time_input, yes_no_prompt
-from .set_date import set_date
 from .validate import (
     BlankInputValidator,
     BooleanValidator,
